+++ NEWS.md
@@ -401,6 +401,9 @@ Library improvements
     This supersedes the old behavior of reinterpret on Arrays. As a result, reinterpreting
     arrays with different alignment requirements (removed in 0.6) is once again allowed ([#23750]).
 
+  * The `keys` of an `Associative` are now an `AbstractSet`. `Base.KeyIterator{<:Associative}`
+    has been changed to `KeySet{K, <:Associative{K}} <: AbstractSet{K}` ([#24580]).
+
 Compiler/Runtime improvements
 -----------------------------